Refactor Redux store and cart logic to /store directory
Moved Redux store, cartSlice, and cartSelector from /contexts to /store for better organization. Updated all relevant imports throughout the codebase to reflect the new paths. Also improved useHome hook to handle authentication loading state before fetching suggestions.
