+++ frontend/src/components/pages/home/ServiceSection.tsx
@@ -5,8 +5,8 @@ import style from "@/styles/base/page.module.css";
 import Image from "next/image";
 import { toast } from "react-toastify";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "@/contexts/store";
-import { addServiceToRoom } from "@/contexts/cartSlice";
+import { RootState } from "@/store/store";
+import { addServiceToRoom } from "@/store/cartSlice";
 import SelectRoomModal from "@/components/modals/SelecteServiceModal";
 import { AnimatedButton } from "@/components/common/Button";
 import { useScrollAnimation } from "@/hooks/useScrollAnimation";
